--- plugin.yml
@@ -1,4 +1,0 @@
-name: AntiExplosion
-main: me.noodles.MainExplosion
-version: 0.3
-author: BGHDNetwork
