+++ src/main/java/me/noodles/antiexplosion/UpdateJoinEvent.java
@@ -1,4 +1,4 @@
-package me.noodles;
+package me.noodles.antiexplosion;
 
 import org.bukkit.event.player.*;
 import org.bukkit.ChatColor;
@@ -14,14 +14,14 @@ public class UpdateJoinEvent implements Listener
     public void onJoin(final PlayerJoinEvent e) {
     	Player p = e.getPlayer();
     	if (p.hasPermission("antiexplosion.update")) {
-    		if (MainExplosion.getPlugin().getConfig().getBoolean("Update.Enabled") == true){
-    		this.checker = new UpdateChecker(MainExplosion.plugin);
+    		if (AntiExplosion.getPlugin().getConfig().getBoolean("Update.Enabled") == true){
+    		this.checker = new UpdateChecker(AntiExplosion.plugin);
                         if (this.checker.isConnected()) {
                             if (this.checker.hasUpdate()) {
                             	p.sendMessage(ChatColor.GRAY + "=========================");
                                 p.sendMessage(ChatColor.RED + "AntiExplosion is outdated!");
                                 p.sendMessage(ChatColor.GREEN + "Newest version: " + this.checker.getLatestVersion());
-                                p.sendMessage(ChatColor.RED + "Your version: " + MainExplosion.plugin.getDescription().getVersion());
+                                p.sendMessage(ChatColor.RED + "Your version: " + AntiExplosion.plugin.getDescription().getVersion());
                                 p.sendMessage(ChatColor.GRAY + "=========================");
                             }
                         }
